requirements.txt
Flask==1.1.4
requests==2.25.1
python-dotenv==0.19.2
